inventoryManagement/requirements.txt
asgiref==3.8.1
autopep8==1.6.0
crispy-bootstrap3==2024.1
crispy-bootstrap5==2024.10
crispy-bulma==0.11.0
dj-database-url==0.5.0
Django==5.1.3
django-crispy-forms==2.3
greenlet==3.1.1
gunicorn==20.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==2.1.3
pandas==2.2.3
Pillow==9.0.1
protobuf==3.20.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0.2
query==0.1.4
setuptools==75.5.0
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
user-agents==2.2.0
whitenoise==6.0.0
